updated saveUser to use validateUser, moved validateCampaign and test
