+++ src/user/validate_campaign.js
@@ -2,6 +2,15 @@ const validCampaigns = [
   'inauguration',
 ]
 
+/**
+ * validateCampaign - validate campaign string
+ * ensures that any campaign given is valid using the list above
+ * if not valid, prefixes it with unknown: so we can do data cleanup on it later
+ * e.g. mispellings, etc.
+ *
+ * @param  {type} campaign = '' description
+ * @return {type}               description
+ */
 module.exports = function validateCampaign(campaign = '') {
   // return false for invalid input
   if (!campaign) return false
